Replace TypeScript build with Rollup bundler
