+++ src/install-hugo.interfaces.ts
@@ -1,0 +1,14 @@
+/**
+ * Install hugo options
+ */
+export interface InstallHugoOptions {
+  arch: string;
+  destination: string;
+  downloadUrl: string;
+  extended: boolean;
+  force: boolean;
+  os: string;
+  skipChecksumCheck: boolean;
+  skipHealthCheck: boolean;
+  version: string;
+}
